requirements.txt
certifi==2022.9.24
elastic-transport==8.4.0
elasticsearch==8.5.2
numpy==1.23.5
python-dotenv==0.21.0
Shapely==1.8.5.post1
urllib3==1.26.13